requirements.txt
openai>=1.3.0
PyPDF2>=3.0.0
pathlib
pandas>=1.5.0
openpyxl>=3.0.0
python-docx>=0.8.11
supabase>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
neo4j>=5.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
